made changes to WarData but there's still a lot of problems
