+++ dbSetup/models/views/__init__.py
@@ -75,13 +75,5 @@ class PitchingStatsView(Base):
     p_FIP = Column(Float, nullable = True) # Fielding Independent Pitching
     playing_time = Column(Float, nullable = True) # Playing Time
 
-    # __table_args__ and relationships not needed because this is a view
-
-class WarData(Base):
-    __tablename__ = "WarData"
-    wardata_ID = Column(Integer, primary_key=True, nullable=False, autoincrement=True)
-    Name = Column(String(255), nullable=False)
-    yearID = Column(SmallInteger, nullable=False)
-    WAR = Column(Float, nullable=True)
-    WRC_Plus = Column(Integer, nullable=True)
+
