+++ dbSetup/services/create_battingstats_view.py
@@ -21,70 +21,62 @@ def create_battingstats_view():
     create_battingstats_view_sql = """
     CREATE OR REPLACE VIEW battingstatsview AS
     SELECT DISTINCT
-        b.playerID as playerID,
-        (b.yearID - p.birthYear) AS age,
-        b.yearID AS yearID,
-        b.teamID AS teamID,
-        b.stint AS stint,
-        p.nameFirst,
-        p.nameLast,
-        p.nameGiven,
-        a.G_ALL AS b_G,
-        (b.b_AB + b.b_BB + b.b_HBP + b.b_SH + b.b_SF) AS b_PA,
-        b.b_HR AS b_HR,
-        b.b_SB AS b_SB,
-        (b.b_BB / (b.b_AB + b.b_BB + b.b_HBP + b.b_SH + b.b_SF)) * 100 AS b_BB_percent,
-        (b.b_SO / (b.b_AB + b.b_BB + b.b_HBP + b.b_SH + b.b_SF)) * 100 AS b_K_percent,
-        (b.b_H - b.b_HR) / (b.b_AB - b.b_SO - b.b_HR + b.b_SF) AS b_BABIP,
-        (b.b_H / b.b_AB) AS b_AVG,
-        ((b.b_H - (b.b_2B + b.b_3B + b.b_HR)) + (2 * b.b_2B) + (3 * b.b_3B) + (4 * b.b_HR)) / b.b_AB AS b_SLG,
-        (((b.b_H - (b.b_2B + b.b_3B + b.b_HR)) + (2 * b.b_2B) + (3 * b.b_3B) + (4 * b.b_HR)) / b.b_AB) - (b.b_H / b.b_AB) AS b_ISO,
-        (b.b_H - (b.b_2B + b.b_3B + b.b_HR)) AS b_b_1B,
-        (b.b_PA / (SELECT SUM(b2.b_PA) 
-               FROM batting b2 
-               WHERE b2.yearID = b.yearID AND b2.teamID = b.teamID)) * 100 AS b_Playing_Time,
-        br.WRC_Plus AS b_wRCPlus,
-                                                -- calculation of 1b
-        (((w.wBB * (b.b_BB - b.b_IBB)) + (w.wHBP * b.b_HBP) + (w.w1b * (b.b_H - (b.b_2B + b.b_3B + b.b_HR))) + (w.w2b * b.b_2B) + (w.w3b * b.b_3B) + (w.whr * b.b_HR))
-            / (b.b_AB + b.b_BB - b.b_IBB + b.b_SF + b.b_HBP))
-            AS b_wOBA,
+    b.playerID as playerID,
+    (b.yearID - p.birthYear) AS age,
+    b.yearID AS yearID,
+    b.teamID AS teamID,
+    b.stint AS stint,
+    p.nameFirst,
+    p.nameLast,
+    p.nameGiven,
+    a.G_ALL AS b_G,
+    b.b_PA,  -- Directly using the b_PA column from the batting table
+    b.b_HR AS b_HR,
+    b.b_SB AS b_SB,
+    (b.b_BB / (b.b_AB + b.b_BB + b.b_HBP + b.b_SH + b.b_SF)) * 100 AS b_BB_percent,
+    (b.b_SO / (b.b_AB + b.b_BB + b.b_HBP + b.b_SH + b.b_SF)) * 100 AS b_K_percent,
+    (b.b_H - b.b_HR) / (b.b_AB - b.b_SO - b.b_HR + b.b_SF) AS b_BABIP,
+    (b.b_H / b.b_AB) AS b_AVG,
+    ((b.b_H - (b.b_2B + b.b_3B + b.b_HR)) + (2 * b.b_2B) + (3 * b.b_3B) + (4 * b.b_HR)) / b.b_AB AS b_SLG,
+    (((b.b_H - (b.b_2B + b.b_3B + b.b_HR)) + (2 * b.b_2B) + (3 * b.b_3B) + (4 * b.b_HR)) / b.b_AB) - (b.b_H / b.b_AB) AS b_ISO,
+    b.b_b_1B,  -- Directly using the b_b_1B column from the batting table
+    (b.b_PA / (SELECT SUM(b2.b_PA) 
+           FROM batting b2 
+           WHERE b2.yearID = b.yearID AND b2.teamID = b.teamID)) * 100 AS b_Playing_Time,
+    br.WRC_Plus AS b_wRCPlus,
+    br.WAR AS b_WAR,
+    -- calculation of wOBA and wRC
+    (((w.wBB * (b.b_BB - b.b_IBB)) + (w.wHBP * b.b_HBP) + (w.w1b * (b.b_H - (b.b_2B + b.b_3B + b.b_HR))) + (w.w2b * b.b_2B) + (w.w3b * b.b_3B) + (w.whr * b.b_HR))
+        / (b.b_AB + b.b_BB - b.b_IBB + b.b_SF + b.b_HBP))
+        AS b_wOBA,
+    (
+    -- woba
+    (    ((((w.wBB * (b.b_BB - b.b_IBB)) + (w.wHBP * b.b_HBP) + (w.w1b * (b.b_H - (b.b_2B + b.b_3B + b.b_HR))) + (w.w2b * b.b_2B) + (w.w3b * b.b_3B) + (w.whr * b.b_HR))
+        / (b.b_AB + b.b_BB - b.b_IBB + b.b_SF + b.b_HBP))
+        - w.league)
+    /
+    w.wobascale)
+    +
+        ( w.r_pa) )
+    *
+    (b.b_AB + b.b_BB + b.b_HBP + b.b_SH + b.b_SF) -- PA
+    AS b_wRC
+FROM
+    batting b
+JOIN
+    people p ON b.playerID = p.playerID
+JOIN
+    appearances a ON b.playerID = a.playerID AND b.yearID = a.yearID AND a.teamID = b.teamID
+JOIN
+    fielding f ON b.playerID = f.playerID AND b.yearID = f.yearID AND a.teamID = b.teamID
+JOIN
+    wobaweights w ON w.yearID = b.yearID
+JOIN 
+    WarData br ON br.playerID = b.playerID AND b.yearID = br.yearID
+WHERE 
+    (b.b_AB + b.b_BB + b.b_HBP + b.b_SH + b.b_SF) > 0  -- exclude pitchers
+    AND (f.position != 'P' OR f.position IS NULL);
 
-        (
-        -- woba
-        (    ((((w.wBB * (b.b_BB - b.b_IBB)) + (w.wHBP * b.b_HBP) + (w.w1b * (b.b_H - (b.b_2B + b.b_3B + b.b_HR))) + (w.w2b * b.b_2B) + (w.w3b * b.b_3B) + (w.whr * b.b_HR))
-            / (b.b_AB + b.b_BB - b.b_IBB + b.b_SF + b.b_HBP))
-            - w.league)
-            /
-            w.wobascale)
-        +
-            ( w.r_pa) )
-        *
-        (b.b_AB + b.b_BB + b.b_HBP + b.b_SH + b.b_SF) -- PA
-        AS b_wRC
-
-        -- ,((b.b_SB - b.b_CS) * 0.2) AS b_BsR
-    FROM
-        batting b
-    JOIN
-        people p ON b.playerID = p.playerID
-    JOIN
-        appearances a ON b.playerID = a.playerID AND b.yearID = a.yearID AND a.teamID = b.teamID
-    JOIN
-        fielding f ON b.playerID = f.playerID AND b.yearID = f.yearID AND a.teamID = b.teamID
-    JOIN
-        wobaweights w ON w.yearID = b.yearID
-
-    JOIN 
-        WarData br ON br.playerID = b.playerID AND b.yearID = br.yearID
-    HAVING
-        b_PA > 0 -- exclude pitchers;
-
-    WHERE 
-        -- exclude pitchers as best we can 
-        (  (b.b_AB + b.b_BB + b.b_HBP + b.b_SH + b.b_SF) > 0
-            OR (b.b_AB + b.b_BB + b.b_HBP + b.b_SH + b.b_SF) IS NULL)
-        AND
-        (f.position != 'P' OR f.position IS NULL);
     """
 
     # Create Batting Stats View
