+++ dbSetup/services/create_pitchingstats_view.py
@@ -90,8 +90,6 @@ def create_pitchingstats_view():
         pi.p_G,
         pi.P_GS,
         pi.p_ERA,
-        (pi.p_IPouts / 3) / (SELECT SUM(p_IPouts) / 3 FROM pitching WHERE yearID = pi.yearID AND teamID = pi.teamID) * 100 AS p_playing_time,
-        CASE
             WHEN pe.birthYear IS NULL THEN
                 NULL -- Age is null if birth year is unknown
             ELSE
